pyproject.toml
[project]
name = "my-own-chatgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
